+++ game/tests/test_views_player.py
@@ -1,6 +1,7 @@
 import random
 
 from django.contrib.auth.models import Permission, User
+from django.core.cache import cache
 from django.core.exceptions import PermissionDenied
 from django.http import Http404
 from django.test import TestCase
@@ -99,6 +100,9 @@ class DiceLaunchViewTest(TestCase):
     def setUp(self):
         self.user = User.objects.last()
         self.client.login(username=self.user.username, password="pwd")
+
+    def tearDown(self):
+        cache.clear()
 
     def test_view_mapping(self):
         game = gmodels.Game.objects.last()
@@ -377,6 +381,9 @@ class ChoiceViewTest(TestCase):
     def setUp(self):
         self.user = User.objects.last()
         self.client.login(username=self.user.username, password="pwd")
+
+    def tearDown(self):
+        cache.clear()
 
     def test_view_mapping(self):
         game = gmodels.Game.objects.last()