+++ role_play/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/", include("django.contrib.auth.urls")),
     path("", RedirectView.as_view(url="game/", permanent=True)),
+    # path("__debug__/", include("debug_toolbar.urls")),
 ]